Prepare for LIFTOFF, Fix injectors to work

+++ Injector/Injector.cpp
@@ -168,9 +168,12 @@ auto main() -> int
 {
 	//VM_START
 	//	STR_ENCRYPTW_START
+
 	unsigned char azukiMagic[] = { 0x61, 0x7a, 0x75, 0x6b, 0x69, 0x5f, 0x6d, 0x61, 0x67, 0x69, 0x63 };
 	unsigned char azukiMagicRev[] = { 0x63, 0x69, 0x67, 0x61, 0x6d, 0x5f, 0x69, 0x6b, 0x75, 0x7a, 0x61 };
-	ShowWindow(GetConsoleWindow(), SW_HIDE);
+	//ShowWindow(GetConsoleWindow(), SW_HIDE);
+
+	// TODO: allocate MemoryRegion for pointer safety
 	auto* mapleBinary = new char[150000000];
 	auto* userData = new char[256 * 5 * 10];
 
@@ -211,9 +214,13 @@ auto main() -> int
 	proc.Attach(osu);
 
 	auto image = proc.mmap().MapImage(oldNtHeader->OptionalHeader.SizeOfImage, mapleBinary, false, blackbone::CreateLdrRef | blackbone::RebaseProcess | blackbone::NoDelayLoad);
+	cacheMemoryRegions(hProcess);
 
-	Sleep(20000);
+	Sleep(10000);
 
+	int timesRedone = 0;
+
+Label_Redo:
 	HMODULE modules[250];
 	DWORD cbNeeded;
 	EnumProcessModules(hProcess, modules, sizeof(modules), &cbNeeded);
@@ -230,6 +237,11 @@ auto main() -> int
 
 	if (ptrUserData == 0 || ptrUserData == nullptr || ptrUserData == NULL)
 	{
+		// Sometimes this scan fails for whatever reason, do the entire thing and try again.
+		if (timesRedone < 3) {
+			timesRedone++;
+			goto Label_Redo;
+		}
 		// MAPLE HAS INJECTED BUT THE SIG SCAN RETURNED ZERO, FUCKING CLOSE OSU
 		// if THIS HERE fails, maple should have an auto process kill if after five seconds, no user data is found within maple
 		TerminateProcess(hProcess, 1);
@@ -238,24 +250,27 @@ auto main() -> int
 	else
 	{
 		// phew, we found what we wanted, good :)
+		std::string userDataString(userData);
 		SIZE_T written = 0;
-		WriteProcessMemory(hProcess, ptrUserData, userData, sizeof userData, &written);
+		WriteProcessMemory(hProcess, ptrUserData, userDataString.c_str(), userDataString.size(), &written);
+		// TODO: rework these checks, can't be bothered right now, let's release maple!!!
+		// 
 		// if we haven't written the entire user-data, kill osu!
-		if (written != sizeof userData)
-		{
-			TerminateProcess(hProcess, 1);
-			CloseHandle(hProcess);
-		}
-		char readBuffer[256 * 2];
-		SIZE_T read = 0;
-		// now one last check :)
-		ReadProcessMemory(hProcess, ptrUserData, &readBuffer, sizeof userData, &read);
+		//if (written != userDataString.size())
+		//{
+		//	TerminateProcess(hProcess, 1);
+		//	CloseHandle(hProcess);
+		//}
+		//char readBuffer[256 * 2];
+		//SIZE_T read = 0;
+		//// now one last check :)
+		//ReadProcessMemory(hProcess, ptrUserData, &readBuffer, sizeof userData, &read);
 
-		if (read != sizeof userData || memcmp(userData, readBuffer, sizeof userData) != 0)
-		{
-			TerminateProcess(hProcess, 1);
-			CloseHandle(hProcess);
-		}
+		//if (read != sizeof userData || memcmp(userData, readBuffer, sizeof userData) != 0)
+		//{
+		//	TerminateProcess(hProcess, 1);
+		//	CloseHandle(hProcess);
+		//}
 
 		// Everything should be handled fine by the injector now, if anything went wrong and we haven't caught it until here, Maple will handle stuff internally aswell
 		// We can sleep now, good night :)
